refactor(discord): move runtime queued placeholder helpers
